refactor: split game page rendering responsibility
